analyzer: move EVMEthAddrFromPreimage to storage
